fastapi-tdd-docker/project/requirements.txt
aiosqlite==0.16.1
asyncpg==0.21.0
click==7.1.2
fastapi==0.62.0
h11==0.12.0
iso8601==0.1.14
pydantic==1.7.3
PyPika==0.44.1
pytz==2020.5
starlette==0.13.6
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
uvicorn==0.13.1
